Feat: 로그인 유저 없다면 post, put, delete 실패

+++ front/src/components/common/ActionSelectorModal.js
@@ -1,8 +1,8 @@
 import React, { useContext, useState } from "react";
 import { ListGroup, Modal } from "react-bootstrap";
 import * as Api from "../../Api";
-import { EditFormContext } from "../../App";
-import ToastWrapper from "./Toast";
+import { EditFormContext, UserStateContext } from "../../App";
+import ToastWrapper from "./ToastWrapper";
 
 /**
  * @param "취소, 수정, 삭제" 등 유저 입력을 받는 모달창
@@ -12,33 +12,41 @@ const ActionSelectorModal = ({
   isActionModalVisible,
   setIsActionModalVisible,
   reviewId,
+  authorId,
   setReviews,
 }) => {
+  const { user: loggedInUser } = useContext(UserStateContext);
   const { setIsEditFormVisible } = useContext(EditFormContext);
   const [toastMsg, setToastMsg] = useState("");
-  //       {/* 결과 토스트 팝업 알림: 삭제되었습니다 */}
 
+  // 결과 토스트 팝업 알림: 삭제 성공 혹은 에러
   const deleteReview = async (reviewId) => {
     if (isActionModalVisible) {
       setIsActionModalVisible(false);
     }
+    if (loggedInUser._id !== authorId) {
+      return setToastMsg("다른사람의 게시물을 삭제할 수 없습니다");
+    }
     try {
       const res = await Api.delete(`reviews/${reviewId}`);
       // to do: error handle
-      if (res) {
-        throw new Error("hi");
-      }
       if (!res.statusText === "OK") throw new Error("서버 에러 받아오기");
       setReviews((current) => {
         return current.filter((review) => review._id !== reviewId);
       });
       setToastMsg("게시물이 삭제되었습니다");
     } catch (err) {
-      console.error(err);
-      // {message, name, code, config, request, response}
-      setToastMsg(err.message);
-      // to do
+      setToastMsg(err);
     }
+  };
+
+  const editReview = () => {
+    if (loggedInUser._id !== authorId) {
+      setIsActionModalVisible(false);
+      return setToastMsg("다른사람의 게시물을 수정할 수 없습니다");
+    }
+    setIsActionModalVisible(false);
+    setIsEditFormVisible(true);
   };
   return (
     <>
@@ -51,14 +59,7 @@ const ActionSelectorModal = ({
         centered
       >
         <ListGroup className="text-center">
-          <ListGroup.Item
-            key="edit"
-            action
-            onClick={() => {
-              setIsActionModalVisible(false);
-              setIsEditFormVisible(true);
-            }}
-          >
+          <ListGroup.Item key="edit" action onClick={editReview}>
             수정
           </ListGroup.Item>
           <ListGroup.Item
@@ -79,7 +80,12 @@ const ActionSelectorModal = ({
         </ListGroup>
       </Modal>
       {toastMsg && (
-        <ToastWrapper onClose={() => setToastMsg("")} text={toastMsg} />
+        <ToastWrapper
+          onClose={() => setToastMsg("")}
+          text={toastMsg}
+          bg="warning"
+          position="bottom-center"
+        />
       )}
     </>
   );
